Optimize travel speeds and filament retraction and extrusion in start and end code.
The filament is extracted in the origin position to place a small dot on the build plate.
From there, a very fine thread of filament will be draws to the object starting position,
which does not interfere with the printing, but assures that there is enough filament in
the print head when the print begins.
On end, the retraction relieves pressure just enough to reduce oozing, but leaves enough
filament in the print head.
Where recalibrating the build plate, remove excess filament from the nozzle, otherwise this
will increase the offset to the build plate surface while recalibrating the cold printer.